- Drafto of refactoring
