+++ mpi4jax/_src/collective_ops/allgather.py
@@ -29,6 +29,8 @@ from ..decorators import (
 from ..validation import enforce_types
 from ..comm import get_default_comm
 
+from ..xla_bridge.device_descriptors import build_allgather_descriptor
+
 # The Jax primitive
 mpi_allgather_p = Primitive("allgather_mpi")  # Create the primitive
 mpi_allgather_impl = default_primitive_impl(mpi_allgather_p)
@@ -182,8 +184,6 @@ def mpi_allgather_xla_encode_device(
 
 @translation_rule_xpu
 def mpi_allgather_xla_encode_xpu(ctx, sendbuf, token, comm):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_allgather_descriptor
-
     return mpi_allgather_xla_encode_device(
         ctx, sendbuf, token, comm, build_allgather_descriptor
     )
@@ -191,8 +191,6 @@ def mpi_allgather_xla_encode_xpu(ctx, sendbuf, token, comm):
 
 @translation_rule_gpu
 def mpi_allgather_xla_encode_gpu(ctx, sendbuf, token, comm):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_allgather_descriptor
-
     return mpi_allgather_xla_encode_device(
         ctx, sendbuf, token, comm, build_allgather_descriptor
     )
