+++ mpi4jax/_src/collective_ops/allreduce.py
@@ -30,6 +30,7 @@ from ..decorators import (
 from ..validation import enforce_types
 from ..comm import get_default_comm
 
+from ..xla_bridge.device_descriptors import build_allreduce_descriptor
 
 # The Jax primitive
 mpi_allreduce_p = Primitive("allreduce_mpi")  # Create the primitive
@@ -176,8 +177,6 @@ def mpi_allreduce_xla_encode_device(
 
 @translation_rule_gpu
 def mpi_allreduce_xla_encode_gpu(ctx, x, token, op, comm, transpose):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_allreduce_descriptor
-
     return mpi_allreduce_xla_encode_device(
         ctx, x, token, op, comm, transpose, build_allreduce_descriptor
     )
@@ -185,8 +184,6 @@ def mpi_allreduce_xla_encode_gpu(ctx, x, token, op, comm, transpose):
 
 @translation_rule_xpu
 def mpi_allreduce_xla_encode_xpu(ctx, x, token, op, comm, transpose):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_allreduce_descriptor
-
     return mpi_allreduce_xla_encode_device(
         ctx, x, token, op, comm, transpose, build_allreduce_descriptor
     )
