+++ mpi4jax/_src/collective_ops/bcast.py
@@ -28,6 +28,7 @@ from ..decorators import (
 )
 from ..validation import enforce_types
 from ..comm import get_default_comm
+from ..xla_bridge.device_descriptors import build_bcast_descriptor
 
 
 # The Jax primitive
@@ -179,8 +180,6 @@ def mpi_bcast_xla_encode_device(ctx, x, token, root, comm, build_bcast_descripto
 
 @translation_rule_xpu
 def mpi_bcast_xla_encode_xpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_bcast_descriptor
-
     return mpi_bcast_xla_encode_device(
         ctx, x, token, root, comm, build_bcast_descriptor
     )
@@ -188,8 +187,6 @@ def mpi_bcast_xla_encode_xpu(ctx, x, token, root, comm):
 
 @translation_rule_gpu
 def mpi_bcast_xla_encode_gpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_bcast_descriptor
-
     return mpi_bcast_xla_encode_device(
         ctx, x, token, root, comm, build_bcast_descriptor
     )
