+++ mpi4jax/_src/collective_ops/gather.py
@@ -28,6 +28,7 @@ from ..decorators import (
 )
 from ..validation import enforce_types
 from ..comm import get_default_comm
+from ..xla_bridge.device_descriptors import build_gather_descriptor
 
 
 # The Jax primitive
@@ -212,8 +213,6 @@ def mpi_gather_xla_encode_device(ctx, x, token, root, comm, build_gather_descrip
 
 @translation_rule_xpu
 def mpi_gather_xla_encode_xpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_gather_descriptor
-
     return mpi_gather_xla_encode_device(
         ctx, x, token, root, comm, build_gather_descriptor
     )
@@ -221,8 +220,6 @@ def mpi_gather_xla_encode_xpu(ctx, x, token, root, comm):
 
 @translation_rule_gpu
 def mpi_gather_xla_encode_gpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_gather_descriptor
-
     return mpi_gather_xla_encode_device(
         ctx, x, token, root, comm, build_gather_descriptor
     )
