+++ mpi4jax/_src/collective_ops/scan.py
@@ -28,6 +28,7 @@ from ..decorators import (
 )
 from ..validation import enforce_types
 from ..comm import get_default_comm
+from ..xla_bridge.device_descriptors import build_scan_descriptor
 
 
 # The Jax primitive
@@ -162,15 +163,11 @@ def mpi_scan_xla_encode_device(ctx, x, token, op, comm, build_scan_descriptor):
 
 @translation_rule_xpu
 def mpi_scan_xla_encode_xpu(ctx, x, token, op, comm):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_scan_descriptor
-
     return mpi_scan_xla_encode_device(ctx, x, token, op, comm, build_scan_descriptor)
 
 
 @translation_rule_gpu
 def mpi_scan_xla_encode_gpu(ctx, x, token, op, comm):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_scan_descriptor
-
     return mpi_scan_xla_encode_device(ctx, x, token, op, comm, build_scan_descriptor)
