+++ mpi4jax/_src/collective_ops/scatter.py
@@ -28,7 +28,7 @@ from ..decorators import (
 )
 from ..validation import enforce_types
 from ..comm import get_default_comm
-
+from ..xla_bridge.device_descriptors import build_scatter_descriptor
 
 # The Jax primitive
 mpi_scatter_p = Primitive("scatter_mpi")  # Create the primitive
@@ -205,8 +205,6 @@ def mpi_scatter_xla_encode_device(ctx, x, token, root, comm, build_scatter_descr
 
 @translation_rule_xpu
 def mpi_scatter_xla_encode_xpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_xpu import build_scatter_descriptor
-
     return mpi_scatter_xla_encode_device(
         ctx, x, token, root, comm, build_scatter_descriptor
     )
@@ -214,8 +212,6 @@ def mpi_scatter_xla_encode_xpu(ctx, x, token, root, comm):
 
 @translation_rule_gpu
 def mpi_scatter_xla_encode_gpu(ctx, x, token, root, comm):
-    from ..xla_bridge.mpi_xla_bridge_gpu import build_scatter_descriptor
-
     return mpi_scatter_xla_encode_device(
         ctx, x, token, root, comm, build_scatter_descriptor
     )
